requirements-viewer.txt
# Requirements for Telegram Archive Web Viewer (standalone)
# This is a lighter dependency set - no Telegram client needed

# Web framework
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.4
httpx>=0.28.0
websockets>=12.0

# Configuration
python-dotenv>=1.0.1

# Database (v3.0+)
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.20.0
asyncpg>=0.30.0
greenlet>=3.1.0

# Web Push notifications (v5.0+)
pywebpush>=2.0.0
py-vapid>=1.9.0

# Image processing for thumbnails (v7.2.0+)
Pillow>=10.0.0






